Added automatic generation of annotated node for all nodes in an investigation after data set parsing.

+++ refinery/data_set_manager/utils.py
@@ -3,11 +3,16 @@ Created on May 29, 2012
 
 @author: nils
 '''
-from data_set_manager.models import Node, Attribute, AnnotatedNode, Study, Assay
+from data_set_manager.models import Node, Attribute, AnnotatedNode, Study, Assay, \
+    AnnotatedNodeRegistry
 from django.db.models import Q
 from django.utils.datetime_safe import datetime
 import simplejson
     
+    
+# number of AnnotatedNode objects that can be inserted with bulk insert (limitation of sqlite)
+# https://docs.djangoproject.com/en/dev/ref/models/querysets/#django.db.models.query.QuerySet.bulk_create
+MAX_BULK_LIST_SIZE = 75
 
 def uniquify(seq):
     set = {}
@@ -243,9 +248,26 @@ def get_matrix( node_type, study_uuid, assay_uuid=None, ontology_attribute_field
     return results
 
 
-def update_annotated_nodes( node_type, study_uuid, assay_uuid=None ):
-    
-    # 1. remove existing annotated node objects for this node_type in this study/assay
+def update_annotated_nodes( node_type, study_uuid, assay_uuid=None, update=False ):
+        
+    # retrieve study and assay ids
+    study = Study.objects.filter( uuid=study_uuid )[0]
+    if assay_uuid is not None:
+        assay = Assay.objects.filter( uuid=assay_uuid )[0]
+    else:
+        assay = None
+            
+    # check if this combination of node_type, study_uuid and assay_uuid already exists
+    if assay is None:
+        registry, created = AnnotatedNodeRegistry.objects.get_or_create( study_id=study.id, assay__isnull=True, node_type=node_type )
+    else:
+        registry, created = AnnotatedNodeRegistry.objects.get_or_create( study_id=study.id, assay_id=assay.id, node_type=node_type )
+        
+    if not created and not update:
+        # registry entry exists and no updating requested  
+        return 
+        
+    # remove existing annotated node objects for this node_type in this study/assay
     if assay_uuid is None:
         counter = AnnotatedNode.objects.filter( Q( study__uuid=study_uuid, assay__uuid__isnull=True ), node_type=node_type ).count()
         AnnotatedNode.objects.filter( Q( study__uuid=study_uuid, assay__uuid__isnull=True ), node_type=node_type ).delete()
@@ -255,50 +277,58 @@ def update_annotated_nodes( node_type, study_uuid, assay_uuid=None ):
     
     print str( counter ) + " annotated nodes deleted."                
     
-    # 2. retrieve annotated nodes
-    nodes = _retrieve_nodes( node_type, study_uuid, assay_uuid, True )
+    # retrieve annotated nodes
+    nodes = _retrieve_nodes( node_type, study_uuid, assay_uuid, True )    
     
-    # 3. retrieve study and assay ids
-    study = Study.objects.filter( uuid=study_uuid )[0]
-    if assay_uuid is not None:
-        assay = Assay.objects.filter( uuid=assay_uuid )[0]
-    else:
-        assay = None    
+    # insert node and attribute information    
+    import time
+    start = time.time()
     
-    # 4. insert node and attribute information
-    #  NON_ONTOLOGY_FIELDS = ["id", "type", "subtype", "value", "value_unit", "node"]
     counter = 0    
+    bulk_list = []
     for node_id, node in nodes.iteritems():
         if node["type"] == node_type:
-            # save attributes (attribute[1], etc. are based on Attribute.ALL_FIELDS 
+            # save attributes (attribute[1], etc. are based on Attribute.ALL_FIELDS) 
             for attribute in _get_parent_attributes( nodes, node_id ):
-                counter += 1                                
-                AnnotatedNode.objects.create( 
-                    #node=Node.objects.filter( id=node["id"] )[0],
-                    node_id=node["id"],
-                    #attribute=Attribute.objects.filter( id=attribute[0] )[0],
-                    attribute_id=attribute[0],
-                    study=study,
-                    assay=assay,
-                    node_uuid=node["uuid"],
-                    node_file=node["file"],
-                    node_type=node["type"],
-                    node_name=node["name"],
-                    attribute_type=attribute[1],
-                    attribute_subtype=attribute[2],
-                    attribute_value=attribute[3],
-                    attribute_value_unit=attribute[4]
-                )
-                    
+                counter += 1
+                
+                bulk_list.append(                                 
+                    AnnotatedNode( 
+                        node_id=node["id"],
+                        attribute_id=attribute[0],
+                        study=study,
+                        assay=assay,
+                        node_uuid=node["uuid"],
+                        node_file=node["file"],
+                        node_type=node["type"],
+                        node_name=node["name"],
+                        attribute_type=attribute[1],
+                        attribute_subtype=attribute[2],
+                        attribute_value=attribute[3],
+                        attribute_value_unit=attribute[4]
+                    ) )
+                
+                if len( bulk_list ) == MAX_BULK_LIST_SIZE:
+                    AnnotatedNode.objects.bulk_create( bulk_list )
+                    bulk_list = []
+    
+    if len( bulk_list ) > 0:
+        AnnotatedNode.objects.bulk_create( bulk_list )
+        bulk_list = []
+    
+    end = time.time()
                  
-    print str( counter ) + " annotated nodes generated."
+    print str( counter ) + " annotated nodes generated in " + str( end - start )
     
-    from django.core import serializers
+    # update registry entry
+    registry.save()
     
-    if assay_uuid is None:
-        result = AnnotatedNode.objects.filter( Q( study_id=study.id, assay__isnull=True ), node_type=node_type )
-    else:
-        result = AnnotatedNode.objects.filter( Q( study_id=study.id, assay__isnull=True ) | Q( study_id=study.id, assay_id=assay.id ), node_type=node_type)
+    #from django.core import serializers
     
-    return ( serializers.serialize("python", result, ensure_ascii=False ) )
+    #if assay_uuid is None:
+    #    result = AnnotatedNode.objects.filter( Q( study_id=study.id, assay__isnull=True ), node_type=node_type )
+    #else:
+    #    result = AnnotatedNode.objects.filter( Q( study_id=study.id, assay__isnull=True ) | Q( study_id=study.id, assay_id=assay.id ), node_type=node_type)
+    
+    #return ( serializers.serialize("python", result, ensure_ascii=False ) )
